+++ tests/integration/transactions/test_trust_set.py
@@ -5,6 +5,7 @@ from tests.integration.it_utils import (
 )
 from tests.integration.reusable_values import WALLET
 from xrpl.models.amounts import IssuedCurrencyAmount
+from xrpl.models.exceptions import XRPLModelException
 from xrpl.models.transactions import TrustSet, TrustSetFlag
 from xrpl.wallet import Wallet
 
@@ -75,3 +76,19 @@ class TestTrustSet(IntegrationTestCase):
             client,
         )
         self.assertTrue(response.is_successful())
+
+        # currency codes must have exactly 3 characters
+        with self.assertRaises(XRPLModelException) as error:
+            TrustSet(
+                account=WALLET.address,
+                flags=TrustSetFlag.TF_SET_NO_RIPPLE,
+                limit_amount=IssuedCurrencyAmount(
+                    issuer=issuer_wallet.address,
+                    currency="abcd",
+                    value="100",
+                ),
+            )
+            self.assertEqual(
+                error.exception.args[0],
+                "{'currency': 'Invalid currency abcd'}",
+            )